Split off NCited into its own file
