+++ src/Nix/Utils.hs
@@ -5,7 +5,10 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TupleSections #-}
+
+{-# OPTIONS_GHC -Wno-missing-signatures #-}
 
 module Nix.Utils (module Nix.Utils, module X) where
 
@@ -15,6 +18,8 @@ import           Control.Monad.Fix
 import qualified Data.Aeson as A
 import qualified Data.Aeson.Encoding as A
 import           Data.Fix
+import           Data.Functor.Compose
+import           Data.Hashable
 import           Data.HashMap.Lazy (HashMap)
 import qualified Data.HashMap.Lazy as M
 import           Data.List (sortOn)
@@ -24,6 +29,7 @@ import qualified Data.Text as Text
 import qualified Data.Vector as V
 import           Lens.Family2 as X
 import           Lens.Family2.Stock (_1, _2)
+import           Lens.Family2.TH
 
 #if ENABLE_TRACING
 import           Debug.Trace as X
@@ -34,6 +40,9 @@ trace = const id
 traceM :: Monad m => String -> m ()
 traceM = const (return ())
 #endif
+
+$(makeLensesBy (\n -> Just ("_" ++ n)) ''Fix)
+$(makeLensesBy (\n -> Just ("_" ++ n)) ''Compose)
 
 type DList a = Endo [a]
 
@@ -124,3 +133,9 @@ uriAwareSplit = go where
                 let ((suffix, _):path) = go (Text.drop 3 e2)
                  in (e1 <> Text.pack "://" <> suffix, PathEntryURI) : path
             | otherwise -> (e1, PathEntryPath) : go (Text.drop 1 e2)
+
+alterF :: (Eq k, Hashable k, Functor f)
+       => (Maybe v -> f (Maybe v)) -> k -> HashMap k v -> f (HashMap k v)
+alterF f k m = f (M.lookup k m) <&> \case
+    Nothing -> M.delete k m
+    Just v  -> M.insert k v m
